fix: resolve circular dependency TDZ in API error parsing
Extract getApiErrorCode, parseApiError, and parseGoogleConnectError to
a new api-error-parsing.util.ts file to break circular dependency with
google.auth.util -> store -> api.util.

This fixes 15 TDZ errors in the test suite.
