+++ packages/web/src/common/apis/util/api-error-parsing.util.ts
@@ -1,0 +1,38 @@
+import { type ZodType } from "zod";
+import {
+  type GoogleConnectErrorResponse,
+  GoogleConnectErrorResponseSchema,
+} from "@core/types/auth.types";
+import {
+  type ApiError,
+  type ApiResponse,
+} from "../api.types";
+
+const getApiErrorData = (error: ApiError): unknown => {
+  return error?.response?.data;
+};
+
+/**
+ * Extracts the error code from an API error's response data.
+ * Returns undefined when the response has no object body with a string `code` property.
+ */
+export const getApiErrorCode = (error: ApiError): string | undefined => {
+  const data = getApiErrorData(error);
+  if (!data || typeof data !== "object" || !("code" in data)) return undefined;
+  const code = (data as { code?: unknown }).code;
+  return typeof code === "string" ? code : undefined;
+};
+
+export const parseApiError = <T>(
+  error: ApiError,
+  schema: ZodType<T>,
+): T | undefined => {
+  const parsed = schema.safeParse(getApiErrorData(error));
+  return parsed.success ? parsed.data : undefined;
+};
+
+export const parseGoogleConnectError = (
+  error: ApiError,
+): GoogleConnectErrorResponse | undefined => {
+  return parseApiError(error, GoogleConnectErrorResponseSchema);
+};
