+++ packages/web/src/common/apis/util/api.util.ts
@@ -1,10 +1,5 @@
-import { type ZodType } from "zod";
 import { GOOGLE_REVOKED } from "@core/constants/sse.constants";
 import { Status } from "@core/errors/status.codes";
-import {
-  type GoogleConnectErrorResponse,
-  GoogleConnectErrorResponseSchema,
-} from "@core/types/auth.types";
 import { handleGoogleRevoked } from "@web/auth/google/util/google.auth.util";
 import { session } from "@web/common/classes/Session";
 import { ENV_WEB } from "../../constants/env.constants";
@@ -20,6 +15,11 @@ import {
   type ApiResponse,
   type SignoutStatus,
 } from "../api.types";
+import {
+  getApiErrorCode,
+  parseApiError,
+  parseGoogleConnectError,
+} from "./api-error-parsing.util";
 
 export const createApiError = (
   config: ApiRequestConfig,
@@ -34,10 +34,6 @@ export const createApiError = (
   return error;
 };
 
-const getApiErrorData = (error: ApiError): unknown => {
-  return error?.response?.data;
-};
-
 export const isApiError = (error: unknown): error is ApiError => {
   return (
     typeof error === "object" &&
@@ -46,30 +42,7 @@ export const isApiError = (error: unknown): error is ApiError => {
   );
 };
 
-/**
- * Extracts the error code from an API error's response data.
- * Returns undefined when the response has no object body with a string `code` property.
- */
-export const getApiErrorCode = (error: ApiError): string | undefined => {
-  const data = getApiErrorData(error);
-  if (!data || typeof data !== "object" || !("code" in data)) return undefined;
-  const code = (data as { code?: unknown }).code;
-  return typeof code === "string" ? code : undefined;
-};
-
-export const parseApiError = <T>(
-  error: ApiError,
-  schema: ZodType<T>,
-): T | undefined => {
-  const parsed = schema.safeParse(getApiErrorData(error));
-  return parsed.success ? parsed.data : undefined;
-};
-
-export const parseGoogleConnectError = (
-  error: ApiError,
-): GoogleConnectErrorResponse | undefined => {
-  return parseApiError(error, GoogleConnectErrorResponseSchema);
-};
+export { getApiErrorCode, parseApiError, parseGoogleConnectError };
 
 export const signOut = async (status: SignoutStatus) => {
   // since there are currently duplicate event fetches,
